Make sure angleconstraints only take angles between 0 and 180 degrees. (#7688)
